requirements.txt
django==3.2
djangorestframework==3.13.1
pandas==1.4.3
matplotlib ==3.6.0
requests==2.28.1
openpyxl == 3.0.10